23-55

+++ php/restoran/kategori/delete.php
@@ -1,13 +1,13 @@
 <?php 
 
-    require_once "../function.php";
+    if (isset($_GET['id'])) {
+        $id = $_GET['id'];
 
-    $id = 16;
+        $sql = "DELETE FROM tblkategori WHERE idkategori = $id";
 
-    $sql = "DELETE FROM tblkategori WHERE idkategori = $id";
+        $db->runSQL($sql);
 
-    $result = mysqli_query($koneksi, $sql);
-
-    header("location:http://localhost/phpsmk/restoran/kategori/select.php");
+        header("location:?f=kategori&m=select");
+    }
 
 ?>
